refactored quite some stuff.
