+++ app/src/main/java/com/m226b/notentool/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.m226_1_notentool;
+package com.m226b.notentool;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
