+++ app/src/main/java/com/m226b/notentool/ProfilePictureLoader.java
@@ -1,4 +1,4 @@
-package com.example.m226_1_notentool;
+package com.m226b.notentool;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -9,20 +9,18 @@ import com.google.firebase.auth.FirebaseUser;
 import java.io.IOException;
 import java.net.URL;
 
-public class ImageLoader implements Runnable {
-    FirebaseAuth mAuth;
-    FirebaseUser user;
-    Bitmap bmp;
+public class ProfilePictureLoader implements Runnable {
+    Bitmap bitmap;
 
     @Override
     public void run() {
-        mAuth = FirebaseAuth.getInstance();
-        user = mAuth.getCurrentUser();
+        FirebaseAuth mAuth = FirebaseAuth.getInstance();
+        FirebaseUser user = mAuth.getCurrentUser();
 
         try {
             if (user != null && user.getPhotoUrl() != null) {
                 URL url = new URL(user.getPhotoUrl().toString());
-                this.bmp = BitmapFactory.decodeStream(url.openConnection().getInputStream());
+                this.bitmap = BitmapFactory.decodeStream(url.openConnection().getInputStream());
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -30,6 +28,6 @@ public class ImageLoader implements Runnable {
     }
 
     public Bitmap getImage() {
-        return bmp;
+        return bitmap;
     }
 }
