+++ app/src/main/res/layout/activity_dashboard.xml
@@ -16,7 +16,7 @@
         app:layout_constraintTop_toTopOf="parent">
 
         <TextView
-            android:id="@+id/appTitle"
+            android:id="@+id/app_name"
             android:layout_width="140dp"
             android:layout_height="wrap_content"
             android:layout_alignParentStart="true"
@@ -25,12 +25,12 @@
             android:layout_marginLeft="40dp"
             android:layout_marginTop="10dp"
             android:layout_marginBottom="10dp"
-            android:text="@string/appTitleText"
+            android:text="@string/app_name"
             android:textColor="#FFF"
             android:textSize="30sp" />
 
         <androidx.cardview.widget.CardView
-            android:id="@+id/circleForProfilePicture"
+            android:id="@+id/circleContainer"
             android:layout_width="50dp"
             android:layout_height="50dp"
             android:layout_alignParentEnd="true"
@@ -45,7 +45,7 @@
                 android:id="@+id/profilePicture"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:contentDescription="@string/profilePictureContentDescription" />
+                android:contentDescription="@string/dashboard_profile_picture_description" />
         </androidx.cardview.widget.CardView>
     </RelativeLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
