+++ app/src/main/res/layout/activity_login.xml
@@ -12,8 +12,8 @@
         android:layout_height="wrap_content"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="100dp"
-        android:textSize="50dp"
-        android:text="NotenTool"/>
+        android:textSize="50sp"
+        android:text="@string/app_name"/>
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/googleSignIn"
