+++ app/src/main/res/layout/activity_profile.xml
@@ -9,7 +9,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Name"
         android:layout_marginTop="15dp"
         android:textAlignment="center"
         android:textSize="20sp"
@@ -19,7 +18,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Email"
         android:layout_below="@+id/name"
         android:layout_marginTop="15dp"
         android:textSize="20sp"
@@ -32,7 +30,7 @@
         android:layout_height="wrap_content"
         android:layout_below="@id/mail"
         android:layout_marginTop="15dp"
-        android:text="logout"
+        android:text="@string/profile_logout"
         android:layout_centerHorizontal="true"
         android:id="@+id/logout" />
 
@@ -43,6 +41,6 @@
         android:layout_below="@id/logout"
         android:layout_marginTop="15dp"
         android:layout_centerHorizontal="true"
-        android:text="@string/dashboardText" />
+        android:text="@string/profile_dashboard" />
 
 </RelativeLayout>
