+++ app/src/test/java/com/m226b/notentool/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.m226_1_notentool;
+package com.m226b.notentool;
 
 import org.junit.Test;
